requirements.txt
tensorflow-gpu==2.0.0
tensorflow-datasets
matplotlib
nltk
